+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add: Utils.monitorFile
 - feat: Utils.readFile and readFileAsync can take a Gio.File
 - improve Button, EventBox hover events
+- parse passed files starting with .
 
 # 1.5.4